requirements.txt
chromadb
google-generativeai
langchain
sentence-transformers
fitz